Rename lmdb_error into mdbx_error
