+++ heed/examples/all-types-poly.rs
@@ -9,12 +9,14 @@ use heed::EnvOpenOptions;
 use serde::{Deserialize, Serialize};
 
 fn main() -> Result<(), Box<dyn Error>> {
-    fs::create_dir_all(Path::new("target").join("zerocopy-dyn.mdb"))?;
+    let path = Path::new("target").join("heed-poly.mdb");
+
+    fs::create_dir_all(&path)?;
 
     let env = EnvOpenOptions::new()
         .map_size(10 * 1024 * 1024 * 1024) // 10GB
         .max_dbs(3000)
-        .open(Path::new("target").join("zerocopy-dyn.mdb"))?;
+        .open(path)?;
 
     // you can specify that a database will support some typed key/data
     //
