+++ heed/examples/multi-env.rs
@@ -1,21 +1,25 @@
 use std::error::Error;
 use std::fs;
+use std::path::Path;
 
 use heed::types::*;
 use heed::{Database, EnvOpenOptions};
 
 fn main() -> Result<(), Box<dyn Error>> {
-    fs::create_dir_all("target/env1.mdb")?;
+    let env1_path = Path::new("target").join("env1.mdb");
+    let env2_path = Path::new("target").join("env2.mdb");
+
+    fs::create_dir_all(&env1_path)?;
     let env1 = EnvOpenOptions::new()
         .map_size(10 * 1024 * 1024 * 1024) // 10GB
         .max_dbs(3000)
-        .open("target/env1.mdb")?;
+        .open(env1_path)?;
 
-    fs::create_dir_all("target/env2.mdb")?;
+    fs::create_dir_all(&env2_path)?;
     let env2 = EnvOpenOptions::new()
         .map_size(10 * 1024 * 1024 * 1024) // 10GB
         .max_dbs(3000)
-        .open("target/env2.mdb")?;
+        .open(env2_path)?;
 
     let db1: Database<Str, ByteSlice> = env1.create_database(Some("hello"))?;
     let db2: Database<OwnedType<u32>, OwnedType<u32>> = env2.create_database(Some("hello"))?;
