+++ heed/src/mdbx_error.rs
